requirements.txt
# ── Web API ───────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0    # ASGI server with WebSocket support
httpx>=0.27.0                # Async HTTP client (used in tests + advisor)
python-multipart>=0.0.9      # multipart/form-data support for FastAPI

# ── Core async runtime ────────────────────────────────────────
aiohttp>=3.9.0
asyncio-throttle>=1.0.0

# ── Data & Numerics ───────────────────────────────────────────
pandas>=2.2.0
numpy>=1.26.0
polars>=0.20.0           # Optional: fast alternative to pandas

# ── Technical Analysis ────────────────────────────────────────
ta>=0.11.0               # Pure-Python TA (TA-Lib alternative)

# ── Data Sources ─────────────────────────────────────────────
akshare>=1.12.0          # Free: A/HK/US market data
tushare>=1.2.89          # Tushare Pro (A-share data, requires token)

# ── Storage ───────────────────────────────────────────────────
redis[asyncio]>=5.0.0
SQLAlchemy[asyncio]>=2.0.0
asyncpg>=0.29.0          # PostgreSQL async driver
psycopg2-binary>=2.9.0   # Sync driver (for Alembic migrations)

# ── Broker SDKs ───────────────────────────────────────────────
futu-api>=9.0.0          # Futu OpenAPI (HK/US/A-share)
# ibapi>=9.81.1          # IBKR TWS API (install separately from IBKR site)

# ── ML ───────────────────────────────────────────────────────
scikit-learn>=1.4.0
lightgbm>=4.3.0
# torch>=2.2.0           # PyTorch (optional, heavy)

# ── NLP / Text Mining ─────────────────────────────────────────
snownlp>=0.12.3          # Chinese sentiment analysis
jieba>=0.42.1            # Chinese word segmentation

# ── Dashboard ────────────────────────────────────────────────
streamlit>=1.32.0

# ── Config & Utils ────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.6.0
tenacity>=8.2.0          # Retry logic

# ── Dev / Testing ─────────────────────────────────────────────
pytest>=8.1.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
black>=24.3.0
ruff>=0.3.0

# ── Browser automation (WebBroker) ────────────────────────────
# After pip install, also run: playwright install chromium
playwright>=1.44.0
